Break lifestream views out into their own modules
Split copy assets, index pages, post pages, archive pages, and hashtag
pages each out into their own modules.

Each module exposes (optionally) the following methods:

- `get`: Return a parsed representation of the database as a
  preprocessing step.
- `prepare`: Modify the database, for example to linkify hashtags.
- `generate`: Generate HTML templates and persist to disk.

Most views use an internal `compile` method for incremenatally compiling
output which is invoked in `generate`.

Each module owns their input and output paths. Each module owns their
page size.

`src/generator/lifestream/index.js` is now an integration point that
orchestrates the build.

This commit improves the parallelism of filesystem operations.
